requirements.txt
# Core
requests>=2.32.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
beautifulsoup4>=4.12.0,<5.0.0
python-dateutil>=2.9.0,<3.0.0
duckduckgo-search>=6.0.0,<7.0.0

# Database
psycopg2-binary>=2.9.9,<3.0.0

# CrewAI
crewai[tools]>=1.14.0,<2.0.0
litellm>=1.80.0,<2.0.0
